Fix stack view layout bug on view appear
When the keyboard was shown in a view, another view's layout is being
changed w.r.t keyboard's bounds. This is because every view is actually
active _at all times_ even though they are not currently being
displayed. For this reason, we solved this by deactivating some layout
constraints and activating some other layout constraints on view appear.