Example layout constraint conflict solving
Normally, you need a certain number of constraints to unambiguosly
layout an element. For example, for a horizontal stack view, these
constraints may be:

  1. Left border of the stack view is aligned with the left border of
superview.
  2. Right border of the stack view is aligned with the right border of
superview.
  3. Stack view is horizontally centered with the superview.
  4. Stack view is vertically centered with the superview.

However, the important point is, these layout constrains are not
actually how they look like. For example, when you say:

  Left border of the stack view is aligned with the left border of
superview.

This actually means:

  The distance between the left border of the stack view and left border
of the superview is C units far.

Usually, this C is set to 0 when you define the constraint. However,
this may not always be the case. In those cases, you need to inspect
each layout constraint which you have defined one by one and fix any
potentially misvalued C values.

Another important point is the following: Even though you have defined
all constraints good, the storyboard may still show layout constraint
conflicts. In that case doing the following may solve the issue:

  1. Select a constraint.
  2. Select the Attributes Inspector of this constraint from the top
left panel..
  3. Uncheck the "Installed" checkbox and check it again.

Doing this is altering something in the storyboard file (which is the
"Main.storyboard" file in this case) and this solves the issue.

Note that "Main.storyboard" is an actual file in XML format, existing in
the project directory. Xcode shows a rendering of this file.